move tests to run outside mixin
