+++ api_tests/preprints/views/test_preprint_actions.py
@@ -50,8 +50,7 @@ class TestPreprintActionFilters(TestReviewActionFilters):
 
         some_rando = AuthUserFactory()
         res = app.get(url, auth=some_rando.auth, expect_errors=True)
-        assert res.status_code == 200
-        assert res.json['data'] == []
+        assert res.status_code == 403
 
 
 @pytest.mark.enable_quickfiles_creation
